(testing): add e2e tests for codelense support
