+++ tests/e2e/specs/basic.e2e.ts
@@ -1,7 +1,7 @@
-import { DefaultTreeItem, Workbench } from 'wdio-vscode-service'
+import { DefaultTreeItem } from 'wdio-vscode-service'
 import { Key } from 'webdriverio'
-import clipboard from 'clipboardy'
 
+import { tryExecuteCommand, clearAllOutputs, getTerminalText } from '../helpers/index.js'
 import { RunmeNotebook } from '../pageobjects/notebook.page.js'
 import { OutputType, StatusBarElements } from '../pageobjects/cell.page.js'
 
@@ -330,54 +330,3 @@ describe('Runme VS Code Extension', async () => {
     })
   })
 })
-
-/**
- * TODO: cannot get text, this is a bug in wdio integration...
- *
- * Replacement for:
- *
- * ```typescript
- * const text = await terminalView.getText()
- * ```
- */
-async function getTerminalText(workbench: Workbench) {
-  const bottomBar = workbench.getBottomBar()
-  await bottomBar.openTerminalView()
-
-  await workbench.executeCommand('Terminal select all')
-  await workbench.executeCommand('Copy')
-  const text = await clipboard.read()
-  await clipboard.write('')
-  return text
-}
-
-async function tryExecuteCommand(workbench: Workbench, command: string) {
-  const cmds = await workbench.openCommandPrompt()
-
-  await cmds.setText(`>${command}`)
-
-  const items = await cmds.getQuickPicks()
-
-  if (items.length > 0 && await items[0].getLabel() !== 'No matching commands') {
-    await cmds.confirm()
-  } else {
-    await cmds.cancel()
-  }
-}
-
-// async function getInputBox(workbench: Workbench) {
-//   if ((await browser.getVSCodeChannel() === 'vscode' && await browser.getVSCodeVersion() >= '1.44.0')
-//     || await browser.getVSCodeVersion() === 'insiders') {
-//     return new InputBox(workbench.locatorMap)
-//   }
-//   return new QuickOpenBox(workbench.locatorMap)
-// }
-
-// async function waitForInputBox(workbench: Workbench) {
-//   return (await getInputBox(workbench)).wait()
-// }
-
-async function clearAllOutputs(workbench: Workbench) {
-  await tryExecuteCommand(workbench, 'Notebook: Clear All Outputs')
-  await tryExecuteCommand(workbench, 'Notebook: Clear Cell Outputs')
-}
